export token-metadata to shared lib
