+++ package.json
@@ -137,6 +137,7 @@
         "@0x/mesh-graphql-client": "^11.0.1",
         "@0x/order-utils": "^10.4.19",
         "@0x/subproviders": "^6.4.2",
+        "@0x/token-metadata": "^0.0.9",
         "@0x/types": "^3.3.2",
         "@0x/utils": "^6.2.1",
         "@0x/web3-wrapper": "^7.4.2",
